Conflicts:
	sources/states/LevelStateManager.cpp

+++ sources/AnimedEntity.cpp
@@ -12,7 +12,7 @@ void switchAnim(AnimatedActor* a) {
 	if (a->frames>=30/(MathUtil::Abs(a->actor.speed))) {
 		RENDERING(a->actor.e)->texture = theRenderingSystem.loadTextureFile(a->anim[a->ind]);
 		a->ind++;
-		if (a->ind==a->anim.size()) a->ind = 0;
+		if (a->ind == (int)a->anim.size()) a->ind = 0;
 		a->frames=0;
 	}
 }
